requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
contourpy==1.1.0
cycler==0.11.0
Cython==3.0.1
dataclasses-json==0.5.9
db-dtypes==1.1.1
decorator==5.1.1
faiss-cpu==1.7.4
flasgger==0.9.7.1
Flask==2.3.2
Flask-Cors==4.0.0
flask-swagger-ui==4.11.1
fonttools==4.41.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.28.0
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
langchain==0.0.230
langchainplus-sdk==0.0.20
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
mdurl==0.1.2
mistune==3.0.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
proto-plus==1.22.3
protobuf==4.23.4
pyarrow
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML
referencing==0.29.1
requests==2.31.0
rich==13.4.2
rpds-py==0.8.11
rsa==4.9
seaborn==0.12.2
Shapely==1.8.5.post1
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.18
streamlit==1.24.1
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
vertexai==0.0.1
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
